app/requirements.txt
h2o-wave
transformers==4.32.1
torch==2.0.0
pillow<10
streamlit==1.26.0
python-dotenv==1.0.0